@@ -16,7 +16,7 @@
 
 buildscript {
     ext {
-        kotlin_version = '1.3.11'
+        kotlin_version = '1.4.21'
     }
     repositories {
         jcenter()
