@@ -1,13 +1,13 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.3.72'
+    ext.kotlin_version = '1.4.21'
     repositories {
         google()
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.6.3'
+        classpath 'com.android.tools.build:gradle:4.1.2'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
     }
 }
@@ -16,8 +16,6 @@ allprojects {
     repositories {
         google()
         jcenter()
-        // TODO: Remove when appcompat:1.3.0 is out of snapshot
-        maven { url 'https://androidx.dev/snapshots/builds/6482645/artifacts/repository' }
     }
 }
 
