@@ -32,7 +32,7 @@ interface ChatRepository {
     fun activateChat(id: Long)
     fun deactivateChat(id: Long)
     fun showAsBubble(id: Long)
-    fun canBubble(): Boolean
+    fun canBubble(id: Long): Boolean
 }
 
 class DefaultChatRepository internal constructor(
@@ -123,12 +123,15 @@ class DefaultChatRepository internal constructor(
 
     override fun updateNotification(id: Long) {
         val chat = chats.getValue(id)
-        notificationHelper.showNotification(chat, false)
+        notificationHelper.showNotification(chat, false, true)
     }
 
     override fun activateChat(id: Long) {
+        val chat = chats.getValue(id)
         currentChat = id
-        notificationHelper.dismissNotification(id)
+        val isPrepopulatedMsgs =
+            chat.messages.size == 2 && chat.messages[0] != null && chat.messages[1] != null
+        notificationHelper.updateNotification(chat, id, isPrepopulatedMsgs)
     }
 
     override fun deactivateChat(id: Long) {
@@ -144,7 +147,8 @@ class DefaultChatRepository internal constructor(
         }
     }
 
-    override fun canBubble(): Boolean {
-        return notificationHelper.canBubble()
+    override fun canBubble(id: Long): Boolean {
+        val chat = chats.getValue(id)
+        return notificationHelper.canBubble(chat.contact)
     }
 }
