@@ -20,6 +20,7 @@ import android.net.Uri
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.map
 import androidx.lifecycle.switchMap
 import com.example.android.people.data.ChatRepository
 import com.example.android.people.data.DefaultChatRepository
@@ -37,8 +38,9 @@ class ChatViewModel @JvmOverloads constructor(
     private var _photoMimeType: String? = null
 
     /**
-     * We want to dismiss a notification when the corresponding chat screen is open. Setting this
-     * to `true` dismisses the current notification and suppresses further notifications.
+     * We want to update the notification when the corresponding chat screen is open. Setting this
+     * to `true` updates the current notification, removing the unread message(s) badge icon and
+     * suppressing further notifications.
      *
      * We do want to keep on showing and updating the notification when the chat screen is opened
      * as an expanded bubble. [ChatFragment] should set this to false if it is launched in
@@ -70,12 +72,7 @@ class ChatViewModel @JvmOverloads constructor(
     /**
      * Whether the "Show as Bubble" button should be shown.
      */
-    val showAsBubbleVisible: LiveData<Boolean> = object : LiveData<Boolean>() {
-        override fun onActive() {
-            // We hide the "Show as Bubble" button if we are not allowed to show the bubble.
-            value = repository.canBubble()
-        }
-    }
+    val showAsBubbleVisible = chatId.map { id -> repository.canBubble(id) }
 
     fun setChatId(id: Long) {
         chatId.value = id
