@@ -20,8 +20,9 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.view.WindowInsets
-import android.view.WindowInsetsAnimation
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsAnimationCompat
+import androidx.core.view.WindowInsetsCompat
 import androidx.fragment.app.Fragment
 import com.google.android.samples.insetsanimation.databinding.FragmentConversationBinding
 
@@ -37,7 +38,7 @@ class ConversationFragment : Fragment() {
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
+    ): View {
         _binding = FragmentConversationBinding.inflate(inflater, container, false)
         return binding.root
     }
@@ -50,21 +51,21 @@ class ConversationFragment : Fragment() {
 
         /**
          * 1) Since our Activity has declared `window.setDecorFitsSystemWindows(false)`, we need to
-         * handle any [WindowInsets] as appropriate.
+         * handle any [WindowInsetsCompat] as appropriate.
          *
          * Our [RootViewDeferringInsetsCallback] will update our attached view's padding to match
-         * the combination of the [WindowInsets.Type.systemBars], and selectively apply the
-         * [WindowInsets.Type.ime] insets, depending on any ongoing WindowInsetAnimations
+         * the combination of the [WindowInsetsCompat.Type.systemBars], and selectively apply the
+         * [WindowInsetsCompat.Type.ime] insets, depending on any ongoing WindowInsetAnimations
          * (see that class for more information).
          */
         val deferringInsetsListener = RootViewDeferringInsetsCallback(
-            persistentInsetTypes = WindowInsets.Type.systemBars(),
-            deferredInsetTypes = WindowInsets.Type.ime()
+            persistentInsetTypes = WindowInsetsCompat.Type.systemBars(),
+            deferredInsetTypes = WindowInsetsCompat.Type.ime()
         )
         // RootViewDeferringInsetsCallback is both an WindowInsetsAnimation.Callback and an
         // OnApplyWindowInsetsListener, so needs to be set as so.
-        binding.root.setWindowInsetsAnimationCallback(deferringInsetsListener)
-        binding.root.setOnApplyWindowInsetsListener(deferringInsetsListener)
+        ViewCompat.setWindowInsetsAnimationCallback(binding.root, deferringInsetsListener)
+        ViewCompat.setOnApplyWindowInsetsListener(binding.root, deferringInsetsListener)
 
         /**
          * 2) The second step is reacting to any animations which run. This can be system driven,
@@ -81,21 +82,23 @@ class ConversationFragment : Fragment() {
          * Note about [TranslateDeferringInsetsAnimationCallback], it relies on the behavior of
          * [RootViewDeferringInsetsCallback] on the layout's root view.
          */
-        binding.messageHolder.setWindowInsetsAnimationCallback(
+        ViewCompat.setWindowInsetsAnimationCallback(
+            binding.messageHolder,
             TranslateDeferringInsetsAnimationCallback(
                 view = binding.messageHolder,
-                persistentInsetTypes = WindowInsets.Type.systemBars(),
-                deferredInsetTypes = WindowInsets.Type.ime(),
+                persistentInsetTypes = WindowInsetsCompat.Type.systemBars(),
+                deferredInsetTypes = WindowInsetsCompat.Type.ime(),
                 // We explicitly allow dispatch to continue down to binding.messageHolder's
                 // child views, so that step 2.5 below receives the call
-                dispatchMode = WindowInsetsAnimation.Callback.DISPATCH_MODE_CONTINUE_ON_SUBTREE
+                dispatchMode = WindowInsetsAnimationCompat.Callback.DISPATCH_MODE_CONTINUE_ON_SUBTREE
             )
         )
-        binding.conversationRecyclerview.setWindowInsetsAnimationCallback(
+        ViewCompat.setWindowInsetsAnimationCallback(
+            binding.conversationRecyclerview,
             TranslateDeferringInsetsAnimationCallback(
                 view = binding.conversationRecyclerview,
-                persistentInsetTypes = WindowInsets.Type.systemBars(),
-                deferredInsetTypes = WindowInsets.Type.ime()
+                persistentInsetTypes = WindowInsetsCompat.Type.systemBars(),
+                deferredInsetTypes = WindowInsetsCompat.Type.ime()
             )
         )
 
@@ -108,11 +111,12 @@ class ConversationFragment : Fragment() {
          * and clear focus for us.
          *
          * Since `binding.messageEdittext` is a child of `binding.messageHolder`, this
-         * [WindowInsetsAnimation.Callback] will only work if the ancestor view's callback uses the
-         * [WindowInsetsAnimation.Callback.DISPATCH_MODE_CONTINUE_ON_SUBTREE] dispatch mode, which
+         * [WindowInsetsAnimationCompat.Callback] will only work if the ancestor view's callback uses the
+         * [WindowInsetsAnimationCompat.Callback.DISPATCH_MODE_CONTINUE_ON_SUBTREE] dispatch mode, which
          * we have done above.
          */
-        binding.messageEdittext.setWindowInsetsAnimationCallback(
+        ViewCompat.setWindowInsetsAnimationCallback(
+            binding.messageEdittext,
             ControlFocusInsetsAnimationCallback(binding.messageEdittext)
         )
 
@@ -130,7 +134,7 @@ class ConversationFragment : Fragment() {
          *
          * Internally, both [InsetsAnimationLinearLayout] & [InsetsAnimationTouchListener] use a
          * class bundled in this sample called [SimpleImeAnimationController], which simplifies
-         * much of the mechanics for controlling a [WindowInsetsAnimation].
+         * much of the mechanics for controlling a [WindowInsetsAnimationCompat].
          */
     }
 
