@@ -17,60 +17,65 @@
 package com.google.android.samples.insetsanimation
 
 import android.view.View
-import android.view.WindowInsets
-import android.view.WindowInsetsAnimation
+import androidx.core.view.OnApplyWindowInsetsListener
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsAnimationCompat
+import androidx.core.view.WindowInsetsCompat
 
 /**
- * A class which extends/implements both [WindowInsetsAnimation.Callback] and
+ * A class which extends/implements both [WindowInsetsAnimationCompat.Callback] and
  * [View.OnApplyWindowInsetsListener], which should be set on the root view in your layout.
  *
  * This class enables the root view is selectively defer handling any insets which match
- * [deferredInsetTypes], to enable better looking [WindowInsetsAnimation]s.
+ * [deferredInsetTypes], to enable better looking [WindowInsetsAnimationCompat]s.
  *
- * An example is the following: when a [WindowInsetsAnimation] is started, the system will dispatch
- * a [WindowInsets] instance which contains the end state of the animation. For the scenario of
+ * An example is the following: when a [WindowInsetsAnimationCompat] is started, the system will dispatch
+ * a [WindowInsetsCompat] instance which contains the end state of the animation. For the scenario of
  * the IME being animated in, that means that the insets contains the IME height. If the view's
  * [View.OnApplyWindowInsetsListener] simply always applied the combination of
- * [WindowInsets.Type.ime] and [WindowInsets.Type.systemBars] using padding, the viewport of any
+ * [WindowInsetsCompat.Type.ime] and [WindowInsetsCompat.Type.systemBars] using padding, the viewport of any
  * child views would then be smaller. This results in us animating a smaller (padded-in) view into
  * a larger viewport. Visually, this results in the views looking clipped.
  *
  * This class allows us to implement a different strategy for the above scenario, by selectively
- * deferring the [WindowInsets.Type.ime] insets until the [WindowInsetsAnimation] is ended.
+ * deferring the [WindowInsetsCompat.Type.ime] insets until the [WindowInsetsAnimationCompat] is ended.
  * For the above example, you would create a [RootViewDeferringInsetsCallback] like so:
  *
  * ```
  * val callback = RootViewDeferringInsetsCallback(
- *     persistentInsetTypes = WindowInsets.Type.systemBars(),
- *     deferredInsetTypes = WindowInsets.Type.ime()
+ *     persistentInsetTypes = WindowInsetsCompat.Type.systemBars(),
+ *     deferredInsetTypes = WindowInsetsCompat.Type.ime()
  * )
  * ```
  *
- * This class is not limited to just IME animations, and can work with any [WindowInsets.Type]s.
+ * This class is not limited to just IME animations, and can work with any [WindowInsetsCompat.Type]s.
  *
  * @param persistentInsetTypes the bitmask of any inset types which should always be handled
  * through padding the attached view
  * @param deferredInsetTypes the bitmask of insets types which should be deferred until after
- * any related [WindowInsetsAnimation]s have ended
+ * any related [WindowInsetsAnimationCompat]s have ended
  */
 class RootViewDeferringInsetsCallback(
     val persistentInsetTypes: Int,
     val deferredInsetTypes: Int
-) : WindowInsetsAnimation.Callback(DISPATCH_MODE_CONTINUE_ON_SUBTREE),
-    View.OnApplyWindowInsetsListener {
+) : WindowInsetsAnimationCompat.Callback(DISPATCH_MODE_CONTINUE_ON_SUBTREE),
+    OnApplyWindowInsetsListener {
     init {
         require(persistentInsetTypes and deferredInsetTypes == 0) {
             "persistentInsetTypes and deferredInsetTypes can not contain any of " +
-                    " same WindowInsets.Type values"
+                    " same WindowInsetsCompat.Type values"
         }
     }
 
     private var view: View? = null
-    private var lastWindowInsets: WindowInsets? = null
+    private var lastWindowInsets: WindowInsetsCompat? = null
 
     private var deferredInsets = false
 
-    override fun onApplyWindowInsets(v: View, windowInsets: WindowInsets): WindowInsets {
+    override fun onApplyWindowInsets(
+        v: View,
+        windowInsets: WindowInsetsCompat
+    ): WindowInsetsCompat {
         // Store the view and insets for us in onEnd() below
         view = v
         lastWindowInsets = windowInsets
@@ -86,30 +91,30 @@ class RootViewDeferringInsetsCallback(
         val typeInsets = windowInsets.getInsets(types)
         v.setPadding(typeInsets.left, typeInsets.top, typeInsets.right, typeInsets.bottom)
 
-        // We return the new WindowInsets.CONSUMED to stop the insets being dispatched any
+        // We return the new WindowInsetsCompat.CONSUMED to stop the insets being dispatched any
         // further into the view hierarchy. This replaces the deprecated
-        // WindowInsets.consumeSystemWindowInsets() and related functions.
-        return WindowInsets.CONSUMED
+        // WindowInsetsCompat.consumeSystemWindowInsets() and related functions.
+        return WindowInsetsCompat.CONSUMED
     }
 
-    override fun onPrepare(animation: WindowInsetsAnimation) {
+    override fun onPrepare(animation: WindowInsetsAnimationCompat) {
         if (animation.typeMask and deferredInsetTypes != 0) {
-            // We defer the WindowInsets.Type.ime() insets if the IME is currently not visible.
-            // This results in only the WindowInsets.Type.systemBars() being applied, allowing
+            // We defer the WindowInsetsCompat.Type.ime() insets if the IME is currently not visible.
+            // This results in only the WindowInsetsCompat.Type.systemBars() being applied, allowing
             // the scrolling view to remain at it's larger size.
             deferredInsets = true
         }
     }
 
     override fun onProgress(
-        insets: WindowInsets,
-        runningAnims: List<WindowInsetsAnimation>
-    ): WindowInsets {
+        insets: WindowInsetsCompat,
+        runningAnims: List<WindowInsetsAnimationCompat>
+    ): WindowInsetsCompat {
         // This is a no-op. We don't actually want to handle any WindowInsetsAnimations
         return insets
     }
 
-    override fun onEnd(animation: WindowInsetsAnimation) {
+    override fun onEnd(animation: WindowInsetsAnimationCompat) {
         if (deferredInsets && (animation.typeMask and deferredInsetTypes) != 0) {
             // If we deferred the IME insets and an IME animation has finished, we need to reset
             // the flag
@@ -119,8 +124,8 @@ class RootViewDeferringInsetsCallback(
             // Ideally we would just call view.requestApplyInsets() and let the normal dispatch
             // cycle happen, but this happens too late resulting in a visual flicker.
             // Instead we manually dispatch the most recent WindowInsets to the view.
-            if (lastWindowInsets != null) {
-                view?.dispatchApplyWindowInsets(lastWindowInsets!!)
+            if (lastWindowInsets != null && view != null) {
+                ViewCompat.dispatchApplyWindowInsets(view!!, lastWindowInsets!!)
             }
         }
     }
