@@ -16,17 +16,17 @@
 
 package com.google.android.samples.insetsanimation
 
-import android.graphics.Insets
 import android.view.View
-import android.view.WindowInsets
-import android.view.WindowInsetsAnimation
+import androidx.core.graphics.Insets
+import androidx.core.view.WindowInsetsAnimationCompat
+import androidx.core.view.WindowInsetsCompat
 
 /**
- * A [WindowInsetsAnimation.Callback] which will translate/move the given view during any
+ * A [WindowInsetsAnimationCompat.Callback] which will translate/move the given view during any
  * inset animations of the given inset type.
  *
  * This class works in tandem with [RootViewDeferringInsetsCallback] to support the deferring of
- * certain [WindowInsets.Type] values during a [WindowInsetsAnimation], provided in
+ * certain [WindowInsetsCompat.Type] values during a [WindowInsetsAnimationCompat], provided in
  * [deferredInsetTypes]. The values passed into this constructor should match those which
  * the [RootViewDeferringInsetsCallback] is created with.
  *
@@ -34,27 +34,27 @@ import android.view.WindowInsetsAnimation
  * @param persistentInsetTypes the bitmask of any inset types which were handled as part of the
  * layout
  * @param deferredInsetTypes the bitmask of insets types which should be deferred until after
- * any [WindowInsetsAnimation]s have ended
+ * any [WindowInsetsAnimationCompat]s have ended
  * @param dispatchMode The dispatch mode for this callback.
- * See [WindowInsetsAnimation.Callback.getDispatchMode].
+ * See [WindowInsetsAnimationCompat.Callback.getDispatchMode].
  */
 class TranslateDeferringInsetsAnimationCallback(
     private val view: View,
     val persistentInsetTypes: Int,
     val deferredInsetTypes: Int,
     dispatchMode: Int = DISPATCH_MODE_STOP
-) : WindowInsetsAnimation.Callback(dispatchMode) {
+) : WindowInsetsAnimationCompat.Callback(dispatchMode) {
     init {
         require(persistentInsetTypes and deferredInsetTypes == 0) {
             "persistentInsetTypes and deferredInsetTypes can not contain any of " +
-                    " same WindowInsets.Type values"
+                    " same WindowInsetsCompat.Type values"
         }
     }
 
     override fun onProgress(
-        insets: WindowInsets,
-        runningAnimations: List<WindowInsetsAnimation>
-    ): WindowInsets {
+        insets: WindowInsetsCompat,
+        runningAnimations: List<WindowInsetsAnimationCompat>
+    ): WindowInsetsCompat {
         // onProgress() is called when any of the running animations progress...
 
         // First we get the insets which are potentially deferred
@@ -76,7 +76,7 @@ class TranslateDeferringInsetsAnimationCallback(
         return insets
     }
 
-    override fun onEnd(animation: WindowInsetsAnimation) {
+    override fun onEnd(animation: WindowInsetsAnimationCompat) {
         // Once the animation has ended, reset the translation values
         view.translationX = 0f
         view.translationY = 0f
