@@ -1,35 +1,31 @@
 /*
+ * Copyright 2020 The Android Open Source Project
  *
- *  * Copyright 2020 The Android Open Source Project
- *  *
- *  * Licensed under the Apache License, Version 2.0 (the "License");
- *  * you may not use this file except in compliance with the License.
- *  * You may obtain a copy of the License at
- *  *
- *  *      http://www.apache.org/licenses/LICENSE-2.0
- *  *
- *  * Unless required by applicable law or agreed to in writing, software
- *  * distributed under the License is distributed on an "AS IS" BASIS,
- *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  * See the License for the specific language governing permissions and
- *  * limitations under the License.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 package com.example.windowmanagersample
 
 import android.graphics.Rect
 import android.view.View
-import android.view.ViewGroup
-import android.view.ViewGroup.MarginLayoutParams
 import android.widget.FrameLayout
 import androidx.window.DisplayFeature
 
 /**
- * Get the bounds of the display feature translated to the View's coordinate space and current
+ * Gets the bounds of the display feature translated to the View's coordinate space and current
  * position in the window. This will also include view padding in the calculations.
  */
-fun getFeatureBoundsInWindow(
+fun getFeaturePositionInViewRect(
     displayFeature: DisplayFeature,
     view: View,
     includePadding: Boolean = true
@@ -54,8 +50,6 @@ fun getFeatureBoundsInWindow(
 
     val featureRectInView = Rect(displayFeature.bounds)
     val intersects = featureRectInView.intersect(viewRect)
-
-    //Checks to see if the display feature overlaps with our view at all
     if ((featureRectInView.width() == 0 && featureRectInView.height() == 0) ||
         !intersects
     ) {
@@ -69,24 +63,16 @@ fun getFeatureBoundsInWindow(
 }
 
 /**
- * Get the layout params for placing a rectangle indicating a display feature inside a
+ * Gets the layout params for placing a rectangle indicating a display feature inside a
  * [FrameLayout].
  */
-fun getLayoutParamsForFeature(displayFeature: DisplayFeature, viewGroup: ViewGroup):
-        MarginLayoutParams? {
-    val featureRectInView = getFeatureBoundsInWindow(displayFeature, viewGroup) ?: return null
+fun getLayoutParamsForFeatureInFrameLayout(displayFeature: DisplayFeature, view: FrameLayout):
+    FrameLayout.LayoutParams? {
+        val featureRectInView = getFeaturePositionInViewRect(displayFeature, view) ?: return null
 
-    val lp = MarginLayoutParams(featureRectInView.width(), featureRectInView.height())
-    lp.leftMargin = featureRectInView.left
-    lp.topMargin = featureRectInView.top
+        val lp = FrameLayout.LayoutParams(featureRectInView.width(), featureRectInView.height())
+        lp.leftMargin = featureRectInView.left
+        lp.topMargin = featureRectInView.top
 
-    // Make sure that zero-wide and zero-high features are still shown
-    if (featureRectInView.left == featureRectInView.right) {
-        lp.width = 1
+        return lp
     }
-    if (featureRectInView.top == featureRectInView.bottom) {
-        lp.height = 1
-    }
-
-    return lp
-}
