@@ -1,26 +1,23 @@
 /*
+ * Copyright 2020 The Android Open Source Project
  *
- *  * Copyright 2020 The Android Open Source Project
- *  *
- *  * Licensed under the Apache License, Version 2.0 (the "License");
- *  * you may not use this file except in compliance with the License.
- *  * You may obtain a copy of the License at
- *  *
- *  *      http://www.apache.org/licenses/LICENSE-2.0
- *  *
- *  * Unless required by applicable law or agreed to in writing, software
- *  * distributed under the License is distributed on an "AS IS" BASIS,
- *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  * See the License for the specific language governing permissions and
- *  * limitations under the License.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 package com.example.windowmanagersample
 
 import android.os.Bundle
 import androidx.core.util.Consumer
-import androidx.core.view.doOnLayout
 import androidx.window.WindowLayoutInfo
 import androidx.window.WindowManager
 import com.example.windowmanagersample.databinding.ActivitySplitLayoutBinding
@@ -30,32 +27,31 @@ class SplitLayoutActivity : BaseSampleActivity() {
 
     private lateinit var windowManager: WindowManager
     private val layoutStateChangeCallback = LayoutStateChangeCallback()
-
     private lateinit var binding: ActivitySplitLayoutBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivitySplitLayoutBinding.inflate(layoutInflater)
         setContentView(binding.root)
-        windowManager = WindowManager(this, getTestBackend())
-        binding.root.doOnLayout {
-            binding.splitLayout.updateWindowLayout(windowManager.windowLayoutInfo)
-        }
+
+        windowManager = getTestBackend()?.let { backend -> WindowManager(this, backend) }
+            ?: WindowManager(this)
     }
 
-    override fun onAttachedToWindow() {
-        super.onAttachedToWindow()
+    override fun onStart() {
+        super.onStart()
         windowManager.registerLayoutChangeCallback(mainThreadExecutor, layoutStateChangeCallback)
     }
 
-    override fun onDetachedFromWindow() {
-        super.onDetachedFromWindow()
+    override fun onStop() {
+        super.onStop()
         windowManager.unregisterLayoutChangeCallback(layoutStateChangeCallback)
     }
 
     inner class LayoutStateChangeCallback : Consumer<WindowLayoutInfo> {
         override fun accept(newLayoutInfo: WindowLayoutInfo) {
-            binding.splitLayout.updateWindowLayout(newLayoutInfo)
+            val splitLayout = binding.splitLayout
+            splitLayout.updateWindowLayout(newLayoutInfo)
         }
     }
 }
