@@ -1,86 +1,102 @@
 /*
+ * Copyright 2020 The Android Open Source Project
  *
- *  * Copyright 2020 The Android Open Source Project
- *  *
- *  * Licensed under the Apache License, Version 2.0 (the "License");
- *  * you may not use this file except in compliance with the License.
- *  * You may obtain a copy of the License at
- *  *
- *  *      http://www.apache.org/licenses/LICENSE-2.0
- *  *
- *  * Unless required by applicable law or agreed to in writing, software
- *  * distributed under the License is distributed on an "AS IS" BASIS,
- *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  * See the License for the specific language governing permissions and
- *  * limitations under the License.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 package com.example.windowmanagersample.backend
 
 import android.app.Activity
 import android.content.Context
-import android.content.ContextWrapper
 import android.graphics.Point
 import android.graphics.Rect
 import androidx.core.util.Consumer
-import androidx.window.*
-import androidx.window.DisplayFeature.TYPE_FOLD
+import androidx.window.DeviceState
+import androidx.window.DisplayFeature
+import androidx.window.FoldingFeature
+import androidx.window.WindowBackend
+import androidx.window.WindowLayoutInfo
+import androidx.window.WindowManager
 import java.util.concurrent.Executor
 
 /**
- * Sample backend implementation that reports a {@link DisplayFeature#TYPE_FOLD} display feature
- * that goes across the screen in the middle. It can be used as a mock backend implementation
- * with [WindowManager] for automated or manual testing.
- * <p> If the application window is portrait-oriented, then the reported fold in the screen will
- * be horizontal (vertically-folding device configuration). If the window is landscape-oriented,
- * then the reported fold will be vertical (horizontally-folding device configuration).
- * @see [WindowManager]
- * @see [getWindowLayoutInfo]
+ * A sample backend that will have a fold in the middle of the screen. The {@link FoldAxis}
+ * specifies which axis is followed. This sample backend can be used to model devices that open
+ * like a clam shell or a book. This is relative to the display's dimensions. For displays that
+ * are taller than they are wide with a short fold axis they will mimic a horizontal fold. For
+ * displays that are wider than they are tall with a short fold axis will mimic a vertical fold.
+ *
+ * <p>The {@link DeviceState} is fixed to have an opened posture. This sample implementation
+ * ignores the device state listener. Combine with {@link InitialValueWindowBackendDecorator} to
+ * receive the initial value through the listener.
+ *
+ * <p>The {@link WindowLayoutInfo} is also fixed to have a {@link TYPE_FOLD} where the longest
+ * dimension runs parallel to the specified {@link FoldAxis}. The fold is placed in the middle
+ * of the display. Like the {@link DeviceState}, the sample implementation ignores the window
+ * layout info listener. Combine with {@link InitialValueWindowBackendDecorator} to
+ * receive the initial value through the listener.
  */
-class MidScreenFoldBackend : WindowBackend {
+class MidScreenFoldBackend(private val foldAxis: FoldAxis) : WindowBackend {
+    /**
+     * The side which the fold axis should be parallel to.
+     */
+    enum class FoldAxis {
+        LONG_DIMENSION,
+        SHORT_DIMENSION
+    }
 
-    private var deviceState = DeviceState.Builder().setPosture(DeviceState.POSTURE_OPENED).build()
-
+    private var deviceState = FoldingFeature.STATE_FLAT
     private var windowLayoutInfoCallback: Consumer<WindowLayoutInfo>? = null
-    private var deviceStateCallback: Consumer<DeviceState>? = null
+    private var windowLayoutInfoExecutor: Executor? = null
 
-    override fun getDeviceState() = deviceState
+    fun getWindowLayoutInfo(context: Context): WindowLayoutInfo {
+        // Use androidx.window.windowMetrics to retrieve current window size
+        val windowMetrics = WindowManager(context).currentWindowMetrics
+        val windowSize = Point(windowMetrics.bounds.width(), windowMetrics.bounds.height())
+        val featureRect = foldRect(windowSize)
 
-    override fun getWindowLayoutInfo(context: Context): WindowLayoutInfo {
-        val activity = getActivityFromContext(context) ?: throw IllegalArgumentException(
-            "Used non-visual Context used with WindowManager. Please use an Activity or a " +
-                    "ContextWrapper around an Activity instead."
-        )
-        val displaySize = Point(activity.window.decorView.width, activity.window.decorView.height)
-        val featureRect = if (displaySize.x >= displaySize.y) { // Landscape
-            Rect(displaySize.x / 2, 0, displaySize.x / 2, displaySize.y)
-        } else { // Portrait
-            Rect(0, displaySize.y / 2, displaySize.x, displaySize.y / 2)
-        }
-
-        val displayFeature = DisplayFeature.Builder()
-            .setBounds(featureRect)
-            .setType(TYPE_FOLD)
-            .build()
+        val displayFeature =
+            FoldingFeature(
+                featureRect,
+                FoldingFeature.TYPE_FOLD,
+                deviceState
+            )
         val featureList = ArrayList<DisplayFeature>()
         featureList.add(displayFeature)
         return WindowLayoutInfo.Builder().setDisplayFeatures(featureList).build()
     }
 
-    /**
-     * Unwrap the hierarchy of [ContextWrapper]-s until [Activity] is reached.
-     * @return Base [Activity] context or `null` if not available.
-     */
-    private fun getActivityFromContext(c: Context?): Activity? {
-        var context = c
-        while (context is ContextWrapper) {
-            if (context is Activity) {
-                return context
-            }
-            context = context.baseContext
+    private fun foldRect(windowSize: Point): Rect {
+        return when (foldAxis) {
+            FoldAxis.LONG_DIMENSION -> longDimensionFold(windowSize)
+            FoldAxis.SHORT_DIMENSION -> shortDimensionFold(windowSize)
         }
-        return null
+    }
+
+    private fun shortDimensionFold(windowSize: Point): Rect {
+        return if (windowSize.x >= windowSize.y) { // Landscape
+            Rect(windowSize.x / 2, 0, windowSize.x / 2, windowSize.y)
+        } else { // Portrait
+            Rect(0, windowSize.y / 2, windowSize.x, windowSize.y / 2)
+        }
+    }
+
+    private fun longDimensionFold(windowSize: Point): Rect {
+        return if (windowSize.x >= windowSize.y) { // Landscape
+            Rect(0, windowSize.y / 2, windowSize.x, windowSize.y / 2)
+        } else { // Portrait
+            Rect(windowSize.x / 2, 0, windowSize.x / 2, windowSize.y)
+        }
     }
 
     /**
@@ -88,23 +104,34 @@ class MidScreenFoldBackend : WindowBackend {
      * [DeviceState.POSTURE_HALF_OPENED].
      * Specific to the type of device we are emulating in this [WindowBackend] implementation
      */
-    fun toggleDeviceHalfOpenedState() {
-        val posture = if (deviceState.posture == DeviceState.POSTURE_OPENED)
-            DeviceState.POSTURE_HALF_OPENED else
-            DeviceState.POSTURE_OPENED
-        deviceState = DeviceState.Builder().setPosture(posture).build()
-        deviceStateCallback?.accept(deviceState)
+    fun toggleDeviceHalfOpenedState(activity: Activity) {
+        deviceState = if (deviceState == FoldingFeature.STATE_FLAT)
+            FoldingFeature.STATE_HALF_OPENED else
+            FoldingFeature.STATE_FLAT
+        windowLayoutInfoExecutor?.execute {
+            windowLayoutInfoCallback?.accept(getWindowLayoutInfo(activity))
+        }
     }
 
+    @Deprecated("Use FoldingFeature to get the state of the hinge instead.")
     override fun registerDeviceStateChangeCallback(
         executor: Executor,
         callback: Consumer<DeviceState>
-    ) {
-        deviceStateCallback = callback
+    ) {}
+
+    @Deprecated("Use FoldingFeature to get the state of the hinge instead.")
+    override fun unregisterDeviceStateChangeCallback(callback: Consumer<DeviceState>) {
     }
 
-    override fun unregisterDeviceStateChangeCallback(callback: Consumer<DeviceState>) {
-        deviceStateCallback = null
+    override fun registerLayoutChangeCallback(
+        activity: Activity,
+        executor: Executor,
+        callback: Consumer<WindowLayoutInfo>
+    ) {
+        windowLayoutInfoCallback = callback
+        windowLayoutInfoExecutor = executor
+
+        executor.execute { callback.accept(getWindowLayoutInfo(activity)) }
     }
 
     override fun registerLayoutChangeCallback(
@@ -112,10 +139,11 @@ class MidScreenFoldBackend : WindowBackend {
         executor: Executor,
         callback: Consumer<WindowLayoutInfo>
     ) {
-        windowLayoutInfoCallback = callback
+        TODO("Not yet implemented")
     }
 
     override fun unregisterLayoutChangeCallback(callback: Consumer<WindowLayoutInfo>) {
         windowLayoutInfoCallback = null
+        windowLayoutInfoExecutor = null
     }
 }
